requirements.txt
streamlit
openai
python-dotenv
pandas
requests
vanna
python-docx
matplotlib
networkx
langchain-core
langgraph
google-search-results
python-pptx